+++ csrc/cuda/nms_iou.h
@@ -25,8 +25,14 @@
 namespace retinanet {
 namespace cuda {
 
+int nms_rotate(int batchSize,
+    const void *const *inputs, void **outputs,
+    size_t count, int detections_per_im, float nms_thresh,
+    void *workspace, size_t workspace_size, cudaStream_t stream);
+
 int iou(
     const void *const *inputs, void **outputs,
     int num_boxes, int num_anchors, cudaStream_t stream);
+
 }
 }